chore: [#248] Move OnlinePeer component into a subcomponent directory
