+++ src/components/OnlinePeersView/OnlinePeer/OnlinePeer.sass
@@ -1,0 +1,2 @@
+.OnlinePeersView
+  .OnlinePeer
