+++ src/components/OnlinePeersView/OnlinePeer/index.js
@@ -1,0 +1,1 @@
+export { default } from './OnlinePeer'
